Work on Scene 1
This is the work thus far on scene one. Was able to get most of the
scene's audio put in and partially cleaned up. Will probably have to
come back in later and add in additional sound effects because there
aren't many right now and it sounds kind of bland. Need to make sure
that people stay interested. Added a filter on to several of them and
have a good radio filter for use now. I also have tried using pink noise
(brownian and white stood out too much) but that also might change at a
later date. The versions of the scene that are in testing will most
likely be completely changed at some point.